GC: use ofm_malloc() and ofm_free() for admin allocation
Using ofm_....() makes it easier to check that memory allocated for GC
administrtation is itself reclaimed.

+++ racket/src/racket/src/mzrt.c
@@ -178,9 +178,9 @@ void *mzrt_thread_stub(void *data){
 }
 
 #ifdef WIN32
-uintptr_t WINAPI mzrt_win_thread_stub(void *data)
+DWORD WINAPI mzrt_win_thread_stub(void *data)
 {
-  return (uintptr_t)mzrt_thread_stub(data);
+  return (DWORD)mzrt_thread_stub(data);
 }
 #endif
